Added Liferay Sales Engineering Action Outcome Evaluator
Refactored the common code to reflect that a Condition Evaluator does not require to update the workflow status.
This is the initial version of the Liferay Sales Engineering Action Outcome Evaluator. It may be it will need to be updated to make use of the workflowExecutionContext to allow different behaviour depending on position within the workflow